feat(deps): Add `serde-bincode-compat` feature to `reth-trie` (#250)
Ref https://github.com/op-rs/op-reth/issues/241

- Fixes comment from deps fix in
https://github.com/op-rs/op-reth/pull/229 (fix correct but comment
misleading)
- `serde-bincode-compat` is a feature which must be enabled when the
`bincode` dependency is used. Asides for dev-deps, this only happens in
`reth-stages`. `reth-exex` runs in `reth-stages`. In this wokrstream we
use the `TrieUpdate` in exex. `serde-bincode-compat` is already
implemented for `TrieUpdates` in `reth-trie-common`, this PR simply
makes that feature accessible via `reth-trie` and in turn
`reth-optimism-trie`, in order to enable it in `reth-optimism-exex`.