fix: exex block lookup
Previously, the ExEx only checked the notification for new blocks,
however if syncing occurs during the backfill job, the ExEx
notification may only contain a subset of the blocks that need to
be executed.

This change fetches each block from the provider instead of relying on
the notification. This should be very quick compared to the work of executing the block and storing it.